+++ tests/test_rle_frag_rea.c
@@ -125,7 +125,7 @@ static int run_test_frag_rea(char *pcap_file_name, uint32_t param_ptype, uint16_
 		/* check Ethernet frame length */
 		if (header.len <= link_len_src || header.len != header.caplen) {
 			PRINT("ERROR Packet #%lu: bad PCAP packet (len = %d, caplen = %d)\n",
-			      test_pcap_counter, header.len, header.caplen);
+			      (long unsigned int)test_pcap_counter, header.len, header.caplen);
 			goto close_input;
 		}
 
@@ -142,7 +142,9 @@ static int run_test_frag_rea(char *pcap_file_name, uint32_t param_ptype, uint16_
 		/* If the proto type is not set, we set it here from the 12th and 13th octets of the Ethernet
 		 * header.. */
 		if (protocol_type == 0) {
-			protocol_type = ntohs(*((uint16_t *)((void *)(packet + ETHER_PTYPE_POS * sizeof(char)))));
+			protocol_type =
+			        ntohs(*((uint16_t *)((void *)(packet + ETHER_PTYPE_POS *
+			                                      sizeof(char)))));
 		}
 		if (rle_transmitter_encap_data(transmitter, in_packet, in_size,
 		                               protocol_type) == C_ERROR) {
@@ -153,8 +155,9 @@ static int run_test_frag_rea(char *pcap_file_name, uint32_t param_ptype, uint16_
 		uint32_t remaining_pdu_size = in_size;
 
 		if (opt_verbose_flag) {
-			PRINT("INFO: PDU number %zu size to send = %zu\n", test_pcap_counter,
-			      in_size);
+			PRINT("INFO: PDU number %zu size to send = %zu\n",
+			      (size_t)test_pcap_counter,
+			      (size_t)in_size);
 		}
 
 		for (;; ) {
@@ -162,7 +165,7 @@ static int run_test_frag_rea(char *pcap_file_name, uint32_t param_ptype, uint16_
 			    (ret_recv == C_REASSEMBLY_OK)) {
 				if (opt_verbose_flag) {
 					PRINT("INFO: Received all fragments of PDU number %zu\n",
-					      test_pcap_counter);
+					      (size_t)test_pcap_counter);
 				}
 				rle_transmitter_free_context(transmitter, nb_frag_id);
 				break;
@@ -273,8 +276,8 @@ close_fake_burst:
 	      "\tTotal size before encapsulation \t%10lu\n"
 	      "\tTotal sent packets \t\t\t%10lu\n"
 	      "\tAverage size per packet \t\t%10.2f\n",
-	      test_pcap_total_sent_size,
-	      test_pcap_counter,
+	      (long unsigned int)test_pcap_total_sent_size,
+	      (long unsigned int)test_pcap_counter,
 	      (float)(test_pcap_total_sent_size / test_pcap_counter));
 
 	print_tx_stats();
